feat: count ender chest items fix: convert doubles to BigDecimals fix: move economy implementation to its own class

+++ common/src/main/java/ovh/mythmc/banco/common/listeners/PlayerJoinListener.java
@@ -4,9 +4,10 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
 import ovh.mythmc.banco.api.Banco;
-import ovh.mythmc.banco.api.economy.Account;
-import ovh.mythmc.banco.api.economy.AccountManager;
+import ovh.mythmc.banco.api.economy.accounts.Account;
+import ovh.mythmc.banco.api.economy.accounts.AccountManager;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class PlayerJoinListener implements Listener {
@@ -19,7 +20,7 @@ public class PlayerJoinListener implements Listener {
         Account account = accountManager.get(uuid);
 
         if (account == null) {
-            account = new Account(uuid, 0, 0);
+            account = new Account(uuid, BigDecimal.valueOf(0), BigDecimal.valueOf(0));
             accountManager.add(account);
         }
